Extract npm and node wrappers into a virtual package.
One thing that is becoming clear is that there is a lot of value in
"inverting" native node APIs from a callback based API, to a
structured effection API. It seems like every single API that we use,
we end up wrapping it with either static functions, or objects that
present a structured API.

This formalizes that approach and wraps node and npm APIs in a set of
"virtual packages" that mirror the set of node APIs

So the `events` package is wrapped by `@effection/events` and node
APIs are wrapped in `@effection/node`:

`fs` -> `@effection/node/fs`
`child_process` -> `@effection/node/child_process`

This follows the technique outlined here
https://dev.to/larswaechter/path-aliases-with-typescript-in-nodejs-4353
to implement the virtual package path aliases with the idea that these
will eventually be extracted into their own NPM packages.

> Note: This doesn't extract the `http` package with this change, but that
> should probably happen at some point.
