+++ package.json
@@ -8,7 +8,8 @@
   "license": "MIT",
   "private": true,
   "scripts": {
-    "start": "ts-node bin/start.ts",
+    "start": "ts-node -r module-alias/register bin/start.ts",
+    "watch": "ts-node -r module-alias/register bin/watch.ts",
     "lint": "eslint src/**/*.ts bin/**/*.ts"
   },
   "devDependencies": {
@@ -22,6 +23,7 @@
     "@types/express": "^4.17.1",
     "@types/node": "^12.7.11",
     "effection": "~0.3.2",
+    "module-alias": "^2.2.2",
     "ts-node": "^8.4.1",
     "typescript": "^3.6.3",
     "websocket": "^1.0.30"
@@ -32,5 +34,8 @@
   "volta": {
     "node": "12.11.1",
     "yarn": "1.19.1"
+  },
+  "_moduleAliases": {
+    "@effection": "src/effection"
   }
 }
